Move methods only used in rest verticle tests to class MODUSERS-309
